requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
eventlet==0.33.3
Flask==2.3.3
Flask-SocketIO==5.3.6
greenlet==3.2.2
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
six==1.17.0
stepic==0.5.0
Werkzeug==3.1.3
wsproto==1.2.0
